fix: 10 logic bugs across all tabs
- Bug 1 (high): Added workspaceDirRef to fix stale closure in event listeners
  so file detection/drag-drop always use the latest configured directory
- Bug 2 (medium): Added searchVersion trigger to reapply filters after
  handleRefreshData, so refresh doesn't bypass active search/filter state
- Bug 3 (medium): Fixed getFormattedName for files without extension
- Bug 4 (medium): Clear previewFile state when deleting a file
- Bug 5 (medium): Removed fake progress delays and false log messages
  from ZipArchiveModal; actual archive_to_zip now drives the flow
- Bug 6 (medium): Added empty targetDir check before backup invocation
- Bug 9 (low): parseInt -> parseFloat for duplicate size group keys

+++ src/components/DuplicateFinder.tsx
@@ -199,7 +199,7 @@ export default function DuplicateFinder({ workspaceDir, onRefreshWorkspace, addL
           <div style={{ display: "flex", flexDirection: "column", gap: "16px" }}>
             {groupKeys.map((groupKey, groupIdx) => {
               const group = duplicateGroups[groupKey];
-              const displayKey = mode === "hash" ? `HASH: ${groupKey.substring(0, 16)}...` : mode === "size" ? `大小: ${formatSize(parseInt(groupKey))}` : `文件名: ${groupKey}`;
+              const displayKey = mode === "hash" ? `HASH: ${groupKey.substring(0, 16)}...` : mode === "size" ? `大小: ${formatSize(parseFloat(groupKey))}` : `文件名: ${groupKey}`;
 
               return (
                 <div
